feat: delete products (owner)

+++ src/controllers/manager/DELETE/deleteProduct.js
@@ -3,14 +3,28 @@ import logger from "../../../utils/logger.js";
 
 export const deleteProduct = async (req, res) => {
   const { productId } = req.body;
+  const user = req.user;
 
   try {
-    const product = await productManager.deleteProduct(productId);
-    logger.info(`Product ${product.title} deleted`);
-    res.redirect("/api/products");
+    const product = await productManager.getProducts(productId);
+
+    if (!product) {
+      return res.status(404).json({ err: "Product not found" });
+    }
+
+    if (user.role === "admin" || product.owner === user.email) {
+      await productManager.deleteProduct(productId);
+      logger.info(`Product ${productId} deleted`);
+      return res.redirect("/api/products");
+    } else {
+      logger.warning(
+        `Permission denied to delete product. ${user.email} is not the owner of the product (${productId}) `
+      );
+      return res.status(403).render("errors/owner-denied");
+    }
   } catch (err) {
     logger.error(`
-      An error occurred while deleting the product from database
+      An error occurred while deleting the product from the database.
       ${err.stack}  
     `);
     res.status(500).json({
